Updated folder names with correct capitalization: - This was required now that TypeScript has been added - When merged with master this may fail as it will likely once more not track letter-casing changes. If this happens we'll add a bug/issue
