--- src/components/HeaderLogin/HeaderLogin.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function HeaderLogin(props) {
-    return (
-        <div>
-            
-        </div>
-    );
-}
-
-export default HeaderLogin;
