--- src/pages/LoginRegister/LoginRegister.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function LoginRegister(props) {
-    return (
-        <div>
-            LOGIN or REGISTER
-        </div>
-    );
-}
-
-export default LoginRegister;
