+++ src/pages/LoginRegister/LoginRegister.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function LoginRegister(props) {
+    return (
+        <div>
+            LOGIN or REGISTER
+        </div>
+    );
+}
+
+export default LoginRegister;
