--- src/pages/PatronPortal/PatronPortal.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function PatronPortal() {
-    return (
-        <div>
-            PATRON PORTAL
-        </div>
-    );
-}
-
-export default PatronPortal;
