+++ src/pages/PatronPortal/PatronPortal.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function PatronPortal() {
+    return (
+        <div>
+            PATRON PORTAL
+        </div>
+    );
+}
+
+export default PatronPortal;
